Add partial code for calculation of moving time
Moving time is the total time when a cyclist is cycling on the bike, not
just sitting. After each revolution of pedals on bicycle (each pulse
from rpm-meter), we set a timeout with predefined threshold (e.g. 2
seconds). If the cyclists makes another revolution (=> pulse), the
timeout is cleared (=> stopped) and a new one is issued. However, if the
cyclist stops cycling for (2 seconds in our case), the timeout will call
its callback function, which dispatches an action that stops calculating
things (=> stops cycling).

The code in this commit has only partial solution. The calculations of
speed, distance, etc still needs to be fixed.