Add resistance level on trip start, end
Cyclist can now select which resistance he/she has on his spin bike. On
each START_TRIP action, a new resistance record is created. When the
cyclist changes the resistance level, the previous resistance record is
modified with the current timestamp and a new resistance record with the
new resistance level is created. On each STOP_TRIP, the last resistance
record is modified with current timestamp.